+++ crates/astria-cli/src/sequencer/account.rs
@@ -121,11 +121,7 @@ impl Nonce {
 #[derive(clap::Args, Debug)]
 struct ArgsInner {
     /// The url of the Sequencer node
-    #[arg(
-        long,
-        env = "SEQUENCER_URL",
-        default_value = crate::DEFAULT_SEQUENCER_RPC
-    )]
+    #[arg(long, env = "SEQUENCER_URL")]
     sequencer_url: String,
     /// The address of the Sequencer account
     address: Address,