+++ crates/astria-conductor/src/sequencer/client.rs
@@ -3,7 +3,7 @@
 use std::time::Duration;
 
 use astria_core::{
-    generated::sequencerblock::v1::{
+    generated::astria::sequencerblock::v1::{
         sequencer_service_client::SequencerServiceClient,
         GetFilteredSequencerBlockRequest,
     },