+++ crates/astria-sequencer-utils/src/genesis_example.rs
@@ -6,6 +6,11 @@ use std::{
 
 use astria_core::{
     generated::{
+        astria::protocol::genesis::v1::{
+            AddressPrefixes,
+            GenesisFees,
+            IbcParameters,
+        },
         connect::{
             marketmap,
             marketmap::v2::{
@@ -18,11 +23,6 @@ use astria_core::{
                 QuotePrice,
             },
             types::v2::CurrencyPair,
-        },
-        protocol::genesis::v1::{
-            AddressPrefixes,
-            GenesisFees,
-            IbcParameters,
         },
     },
     primitive::v1::Address,
@@ -160,8 +160,8 @@ fn address_prefixes() -> AddressPrefixes {
 }
 
 #[expect(clippy::too_many_lines, reason = "all lines reasonably necessary")]
-fn proto_genesis_state() -> astria_core::generated::protocol::genesis::v1::GenesisAppState {
-    astria_core::generated::protocol::genesis::v1::GenesisAppState {
+fn proto_genesis_state() -> astria_core::generated::astria::protocol::genesis::v1::GenesisAppState {
+    astria_core::generated::astria::protocol::genesis::v1::GenesisAppState {
         accounts: accounts().into_iter().map(Protobuf::into_raw).collect(),
         address_prefixes: Some(address_prefixes()),
         authority_sudo_address: Some(alice().to_raw()),