+++ crates/astria-sequencer/src/action_handler/impls/ics20_withdrawal.rs
@@ -6,7 +6,9 @@ use astria_core::{
     },
     protocol::{
         memos::v1::Ics20WithdrawalFromRollup,
-        transaction::v1::action,
+        transaction::v1::action::{
+            self,
+        },
     },
 };
 use astria_eyre::{
@@ -19,10 +21,12 @@ use astria_eyre::{
         WrapErr as _,
     },
 };
+use async_trait::async_trait;
 use cnidarium::{
     StateRead,
     StateWrite,
 };
+use ibc_proto::ibc::apps::transfer::v2::FungibleTokenPacketData;
 use ibc_types::core::channel::{
     ChannelId,
     PortId,
@@ -33,21 +37,18 @@ use penumbra_ibc::component::packet::{
     SendPacketWrite as _,
     Unchecked,
 };
-use penumbra_proto::core::component::ibc::v1::FungibleTokenPacketData;
 
 use crate::{
     accounts::{
-        AddressBytes as _,
+        AddressBytes,
         StateWriteExt as _,
     },
+    action_handler::ActionHandler,
     address::StateReadExt as _,
-    app::{
-        ActionHandler,
-        StateReadExt as _,
-    },
+    app::StateReadExt as _,
     bridge::{
         StateReadExt as _,
-        StateWriteExt as _,
+        StateWriteExt,
     },
     ibc::{
         StateReadExt as _,
@@ -56,92 +57,7 @@ use crate::{
     transaction::StateReadExt as _,
 };
 
-async fn create_ibc_packet_from_withdrawal<S: StateRead>(
-    withdrawal: &action::Ics20Withdrawal,
-    state: S,
-) -> Result<IBCPacket<Unchecked>> {
-    let sender = if withdrawal.use_compat_address {
-        let ibc_compat_prefix = state.get_ibc_compat_prefix().await.context(
-            "need to construct bech32 compatible address for IBC communication but failed reading \
-             required prefix from state",
-        )?;
-        withdrawal
-            .return_address()
-            .to_prefix(&ibc_compat_prefix)
-            .context("failed to convert the address to the bech32 compatible prefix")?
-            .to_format::<Bech32>()
-            .to_string()
-    } else {
-        withdrawal.return_address.to_string()
-    };
-    let packet = FungibleTokenPacketData {
-        amount: withdrawal.amount.to_string(),
-        denom: withdrawal.denom.to_string(),
-        sender,
-        receiver: withdrawal.destination_chain_address.clone(),
-        memo: withdrawal.memo.clone(),
-    };
-
-    let serialized_packet_data =
-        serde_json::to_vec(&packet).context("failed to serialize fungible token packet as JSON")?;
-
-    Ok(IBCPacket::new(
-        PortId::transfer(),
-        withdrawal.source_channel().clone(),
-        *withdrawal.timeout_height(),
-        withdrawal.timeout_time(),
-        serialized_packet_data,
-    ))
-}
-
-/// Establishes the withdrawal target.
-///
-/// The function returns the following addresses under the following conditions:
-/// 1. `action.bridge_address` if `action.bridge_address` is set and `from` is its stored withdrawer
-///    address.
-/// 2. `from` if `action.bridge_address` is unset and `from` is *not* a bridge account.
-///
-/// Errors if:
-/// 1. Errors reading from DB
-/// 2. `action.bridge_address` is set, but `from` is not the withdrawer address.
-/// 3. `action.bridge_address` is unset, but `from` is a bridge account.
-async fn establish_withdrawal_target<'a, S: StateRead>(
-    action: &'a action::Ics20Withdrawal,
-    state: &S,
-    from: &'a [u8; 20],
-) -> Result<&'a [u8; 20]> {
-    // If the bridge address is set, the withdrawer on that address must match
-    // the from address.
-    if let Some(bridge_address) = &action.bridge_address {
-        let Some(withdrawer) = state
-            .get_bridge_account_withdrawer_address(bridge_address)
-            .await
-            .wrap_err("failed to get bridge withdrawer")?
-        else {
-            bail!("bridge address must have a withdrawer address set");
-        };
-
-        ensure!(
-            &withdrawer == from.address_bytes(),
-            "sender does not match bridge withdrawer address; unauthorized"
-        );
-
-        return Ok(bridge_address.as_bytes());
-    }
-
-    // If the bridge address is not set, the sender must not be a bridge account.
-    if state
-        .is_a_bridge_account(from)
-        .await
-        .context("failed to establish whether the sender is a bridge account")?
-    {
-        bail!("sender cannot be a bridge address if bridge address is not set");
-    }
-
-    Ok(from)
-}
-
-#[async_trait::async_trait]
+#[async_trait]
 impl ActionHandler for action::Ics20Withdrawal {
     // TODO(https://github.com/astriaorg/astria/issues/1430): move checks to the `Ics20Withdrawal` parsing.
     async fn check_stateless(&self) -> Result<()> {
@@ -257,6 +173,91 @@ impl ActionHandler for action::Ics20Withdrawal {
     }
 }
 
+async fn create_ibc_packet_from_withdrawal<S: StateRead>(
+    withdrawal: &action::Ics20Withdrawal,
+    state: S,
+) -> Result<IBCPacket<Unchecked>> {
+    let sender = if withdrawal.use_compat_address {
+        let ibc_compat_prefix = state.get_ibc_compat_prefix().await.context(
+            "need to construct bech32 compatible address for IBC communication but failed reading \
+             required prefix from state",
+        )?;
+        withdrawal
+            .return_address()
+            .to_prefix(&ibc_compat_prefix)
+            .context("failed to convert the address to the bech32 compatible prefix")?
+            .to_format::<Bech32>()
+            .to_string()
+    } else {
+        withdrawal.return_address.to_string()
+    };
+    let packet = FungibleTokenPacketData {
+        amount: withdrawal.amount.to_string(),
+        denom: withdrawal.denom.to_string(),
+        sender,
+        receiver: withdrawal.destination_chain_address.clone(),
+        memo: withdrawal.memo.clone(),
+    };
+
+    let serialized_packet_data =
+        serde_json::to_vec(&packet).context("failed to serialize fungible token packet as JSON")?;
+
+    Ok(IBCPacket::new(
+        PortId::transfer(),
+        withdrawal.source_channel().clone(),
+        *withdrawal.timeout_height(),
+        withdrawal.timeout_time(),
+        serialized_packet_data,
+    ))
+}
+
+/// Establishes the withdrawal target.
+///
+/// The function returns the following addresses under the following conditions:
+/// 1. `action.bridge_address` if `action.bridge_address` is set and `from` is its stored withdrawer
+///    address.
+/// 2. `from` if `action.bridge_address` is unset and `from` is *not* a bridge account.
+///
+/// Errors if:
+/// 1. Errors reading from DB
+/// 2. `action.bridge_address` is set, but `from` is not the withdrawer address.
+/// 3. `action.bridge_address` is unset, but `from` is a bridge account.
+async fn establish_withdrawal_target<'a, S: StateRead>(
+    action: &'a action::Ics20Withdrawal,
+    state: &S,
+    from: &'a [u8; 20],
+) -> Result<&'a [u8; 20]> {
+    // If the bridge address is set, the withdrawer on that address must match
+    // the from address.
+    if let Some(bridge_address) = &action.bridge_address {
+        let Some(withdrawer) = state
+            .get_bridge_account_withdrawer_address(bridge_address)
+            .await
+            .wrap_err("failed to get bridge withdrawer")?
+        else {
+            bail!("bridge address must have a withdrawer address set");
+        };
+
+        ensure!(
+            &withdrawer == from.address_bytes(),
+            "sender does not match bridge withdrawer address; unauthorized"
+        );
+
+        return Ok(bridge_address.as_bytes());
+    }
+
+    // If the bridge address is not set, the sender must not be a bridge account.
+    if state
+        .is_a_bridge_account(from)
+        .await
+        .context("failed to establish whether the sender is a bridge account")?
+    {
+        bail!("sender cannot be a bridge address if bridge address is not set");
+    }
+
+    Ok(from)
+}
+
 fn is_source(source_port: &PortId, source_channel: &ChannelId, asset: &Denom) -> bool {
     if let Denom::TracePrefixed(trace) = asset {
         !trace.has_leading_port(source_port) || !trace.has_leading_channel(source_channel)
@@ -267,27 +268,34 @@ fn is_source(source_port: &PortId, source_channel: &ChannelId, asset: &Denom) ->
 
 #[cfg(test)]
 mod tests {
-    use astria_core::primitive::v1::RollupId;
+    use astria_core::{
+        primitive::v1::RollupId,
+        protocol::transaction::v1::action,
+    };
     use cnidarium::StateDelta;
     use ibc_types::core::client::Height;
 
-    use super::*;
     use crate::{
+        action_handler::impls::{
+            ics20_withdrawal::establish_withdrawal_target,
+            test_utils::test_asset,
+        },
         address::StateWriteExt as _,
         benchmark_and_test_utils::{
             assert_eyre_error,
             astria_address,
             ASTRIA_PREFIX,
         },
+        bridge::StateWriteExt as _,
     };
 
     #[tokio::test]
-    async fn sender_is_withdrawal_target_if_bridge_is_not_set_and_sender_is_not_bridge() {
+    async fn withdrawal_target_is_sender_if_bridge_is_not_set_and_sender_is_not_bridge() {
         let storage = cnidarium::TempStorage::new().await.unwrap();
         let snapshot = storage.latest_snapshot();
         let state = StateDelta::new(snapshot);
 
-        let denom = "test".parse::<Denom>().unwrap();
+        let denom = test_asset();
         let from = [1u8; 20];
         let action = action::Ics20Withdrawal {
             amount: 1,
@@ -312,7 +320,7 @@ mod tests {
     }
 
     #[tokio::test]
-    async fn sender_is_withdrawal_target_if_bridge_is_unset_but_sender_is_bridge() {
+    async fn withdrawal_target_is_sender_if_bridge_is_unset_but_sender_is_bridge() {
         let storage = cnidarium::TempStorage::new().await.unwrap();
         let snapshot = storage.latest_snapshot();
         let mut state = StateDelta::new(snapshot);
@@ -331,7 +339,7 @@ mod tests {
             .put_bridge_account_withdrawer_address(&bridge_address, bridge_address)
             .unwrap();
 
-        let denom = "test".parse::<Denom>().unwrap();
+        let denom = test_asset();
         let action = action::Ics20Withdrawal {
             amount: 1,
             denom: denom.clone(),
@@ -361,21 +369,17 @@ mod tests {
             [1; 20]
         }
 
-        fn denom() -> Denom {
-            "test".parse().unwrap()
-        }
-
         fn action() -> action::Ics20Withdrawal {
             action::Ics20Withdrawal {
                 amount: 1,
-                denom: denom(),
+                denom: test_asset(),
                 bridge_address: None,
                 destination_chain_address: "test".to_string(),
                 return_address: astria_address(&[1; 20]),
                 timeout_height: Height::new(1, 1).unwrap(),
                 timeout_time: 1,
                 source_channel: "channel-0".to_string().parse().unwrap(),
-                fee_asset: denom(),
+                fee_asset: test_asset(),
                 memo: String::new(),
                 use_compat_address: false,
             }
@@ -439,7 +443,7 @@ mod tests {
             .put_bridge_account_withdrawer_address(&bridge_address, withdrawer_address)
             .unwrap();
 
-        let denom = "test".parse::<Denom>().unwrap();
+        let denom = test_asset();
         let action = action::Ics20Withdrawal {
             amount: 1,
             denom: denom.clone(),
@@ -471,7 +475,7 @@ mod tests {
         // sender is not the withdrawer address, so must fail
         let not_bridge_address = [1u8; 20];
 
-        let denom = "test".parse::<Denom>().unwrap();
+        let denom = test_asset();
         let action = action::Ics20Withdrawal {
             amount: 1,
             denom: denom.clone(),