+++ crates/astria-sequencer/src/connect/marketmap/mod.rs
@@ -1,2 +1,0 @@
-pub(crate) mod component;
-pub(crate) mod state_ext;