+++ crates/astria-sequencer/src/service/info/abci_query_router.rs
@@ -40,7 +40,6 @@ use std::{
 
 use cnidarium::Storage;
 use matchit::{
-    InsertError,
     Match,
     MatchError,
 };
@@ -48,6 +47,13 @@ use tendermint::abci::{
     request,
     response,
 };
+
+#[derive(Debug, thiserror::Error)]
+#[error("`{route}` is an invalid route")]
+pub(crate) struct InsertError {
+    route: String,
+    source: matchit::InsertError,
+}
 
 /// `Router` is a wrapper around [`matchit::Router`] to route abci queries
 /// to handlers.
@@ -75,8 +81,13 @@ impl Router {
         route: impl Into<String>,
         handler: impl AbciQueryHandler,
     ) -> Result<(), InsertError> {
+        let route = route.into();
         self.query_router
-            .insert(route, BoxedAbciQueryHandler::from_handler(handler))
+            .insert(route.clone(), BoxedAbciQueryHandler::from_handler(handler))
+            .map_err(|source| InsertError {
+                route,
+                source,
+            })
     }
 }