+++ tools/protobuf-compiler/src/main.rs
@@ -1,3 +1,8 @@
+//! Generates Rust code of protobuf specs located in proto/ and writes
+//! the result to crates/astria-core/src/generated.
+//!
+//! This tool will delete everything in crates/astria-core/src/generated (except
+//! mod.rs).
 use std::{
     collections::{
         HashMap,
@@ -59,6 +64,8 @@ fn main() {
 
     let files = find_protos(src_dir);
 
+    purge_out_dir(&out_dir);
+
     tonic_build::configure()
         .build_client(true)
         .build_server(true)
@@ -67,7 +74,6 @@ fn main() {
         .bytes([
             ".astria",
             ".astria_vendored.tendermint.abci",
-            ".connect",
             ".celestia",
             ".connect",
             ".cosmos",
@@ -209,14 +215,38 @@ fn get_buf_from_env() -> PathBuf {
         "linux" => "You can download it from https://github.com/bufbuild/buf/releases; if you are on Arch Linux, install it from the AUR with `rua install buf` or another helper",
         _other =>  "Check if there is a precompiled version for your OS at https://github.com/bufbuild/buf/releases"
     };
-    let error_msg = "Could not find `buf` installation and this build crate cannot proceed without
-    this knowledge. If `buf` is installed and this crate had trouble finding
-    it, you can set the `BUF` environment variable with the specific path to your
-    installed `buf` binary.";
+    let error_msg = "Could not find `buf` installation and this build crate cannot proceed \
+                     without this knowledge. If `buf` is installed and this crate had trouble \
+                     finding it, you can set the `BUF` environment variable with the specific \
+                     path to your installed `buf` binary.";
     let msg = format!("{error_msg} {os_specific_hint}");
 
     env::var_os("BUF")
         .map(PathBuf::from)
         .or_else(|| which::which("buf").ok())
         .expect(&msg)
+}
+
+fn purge_out_dir(path: impl AsRef<Path>) {
+    let read_dir_msg = format!(
+        "should be able to read generated file out dir files `{}`",
+        path.as_ref().display()
+    );
+    let entry_msg = format!(
+        "every entry in generated file out dir `{}` should have a name",
+        path.as_ref().display()
+    );
+    let remove_msg = format!(
+        "all entries in the generated file out dir should `{}` should be files, and the out dir \
+         is expected to have read, write, execute permissions set",
+        path.as_ref().display()
+    );
+    for entry in read_dir(path).expect(&read_dir_msg).flatten() {
+        // skip mod.rs as it's assumed to be the only non-generated file in the out dir.
+        if entry.path().file_name().expect(&entry_msg) == "mod.rs" {
+            continue;
+        }
+
+        std::fs::remove_file(entry.path()).expect(&remove_msg);
+    }
 }