+++ crates/astria-cli/CHANGELOG.md
@@ -18,6 +18,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fixed ICS20 withdrawal source when using channel with more than one
   port/channel combo. [#1768](https://github.com/astriaorg/astria/pull/1768)
 
+### Changed
+
+- Removed default values from `--sequencer.chain-id` and `--sequencer-url` arguments
+  [#1792](https://github.com/astriaorg/astria/pull/1792)
+
 ## [0.5.1] - 2024-10-23
 
 ### Added